move common input components to invoices/
