+++ src/screens/invoices/components/address-input-group.tsx
@@ -1,6 +1,6 @@
 import { Building2Icon, BuildingIcon, HashIcon, MapPinIcon } from "lucide-react";
-import InputGroup from "../input-group";
-import TextInput from "../text-input";
+import InputGroup from "../../../components/input/input-group";
+import TextInput from "../../../components/input/text-input";
 
 export default function AddressInputGroup() {
   return (
